requirements.txt
alembic==1.16.4
altair==4.2.2
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
docker==7.1.0
entrypoints==0.4
executing==2.2.0
fastapi==0.115.0
Flask==3.1.1
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
gradio==3.36.1
gradio-client==0.2.10
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==22.0.0
h11==0.16.0
idna==3.10
importlib_metadata==7.2.1
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lightgbm==4.6.0
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.3
mlflow==2.14.1
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-gcp-trace
optuna==4.4.0
pandas==2.1.4
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
posthog==6.5.0
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==27.0.1
querystring-parser==1.2.4
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.2
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.38.6
statsmodels==0.14.5
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.30.5
wcwidth==0.2.13
Werkzeug==3.1.3
xgboost==3.0.3
zipp==3.23.0